+++ lib/regexp_parser/expression.rb
@@ -25,6 +25,7 @@ require 'regexp_parser/expression/classes/root'
 require 'regexp_parser/expression/classes/unicode_property'
 
 require 'regexp_parser/expression/methods/construct'
+require 'regexp_parser/expression/methods/human_name'
 require 'regexp_parser/expression/methods/match'
 require 'regexp_parser/expression/methods/match_length'
 require 'regexp_parser/expression/methods/options'